Updated Documentation and New Functions
I have updated the documentation for the functions, including the arguments, so that the input will now be "country" rather than "a". I have added two new functions: read_data reads and downloads the dataset directly from the github repository to the global environment of the user, and discrimination_law provides information on whether laws prevent discrimination in different countries.